+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/impl/RobotPythonExpressionBodyImpl.java
@@ -10,39 +10,21 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotTypes.*;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.*;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotPsiUtil;
-import com.intellij.openapi.util.TextRange;
 
-public class RobotConditionalContentImpl extends RobotPsiElementBase implements RobotConditionalContent {
+public class RobotPythonExpressionBodyImpl extends RobotPsiElementBase implements RobotPythonExpressionBody {
 
-  public RobotConditionalContentImpl(@NotNull ASTNode node) {
+  public RobotPythonExpressionBodyImpl(@NotNull ASTNode node) {
     super(node);
   }
 
   public void accept(@NotNull RobotVisitor visitor) {
-    visitor.visitConditionalContent(this);
+    visitor.visitPythonExpressionBody(this);
   }
 
   @Override
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof RobotVisitor) accept((RobotVisitor)visitor);
     else super.accept(visitor);
-  }
-
-  @Override
-  @NotNull
-  public List<RobotPythonExpressionBody> getPythonExpressionBodyList() {
-    return PsiTreeUtil.getChildrenOfTypeAsList(this, RobotPythonExpressionBody.class);
-  }
-
-  @Override
-  @NotNull
-  public List<RobotVariable> getVariableList() {
-    return PsiTreeUtil.getChildrenOfTypeAsList(this, RobotVariable.class);
-  }
-
-  @Override
-  public @NotNull TextRange getInjectionRelevantTextRange() {
-    return RobotPsiUtil.getInjectionRelevantTextRange(this);
   }
 
 }