+++ CHANGELOG.md
@@ -6,9 +6,14 @@
 
 ### Added
 
+- Language injection support for python expressions in Robot files
+
 ### Changed
 
 ### Fixed
+
+- Don't mark variables containing decimal numbers with a dot as variables missing a variable definition
+- Formatting rules
 
 ## 0.7.1 - 2026-05-02